fixed custom card
